third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFETileElement-svgdom-in-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in' property of the SVGFETileElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tile.in1.baseVal is "SourceGraphic"
PASS tile.in1.baseVal is "offset"
PASS successfullyParsed is true

TEST COMPLETE
